Initial commit
The result is surprising, I expected to see more variation in thread
execution, but it seems to do a bunch of work in one thread before
another thread has chance to run.  My current suspicion without doing
any futher digging in is just that because it's not a real world example
the lock is held for the majority of the "work" so other threads are
just waiting on the mutex.  I could probably achieve what I was looking
at by doing some extra work after the mutex unlocks.